requirements-transcribe.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
backports-datetime-fromisoformat==2.0.3
braceexpand==0.1.7
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
Cython==3.0.12
cytoolz==1.0.1
datasets==2.14.4
decorator==5.2.1
dill==0.3.7
Distance==0.1.3
docker-pycreds==0.4.0
docopt==0.6.2
editdistance==0.8.1
einops==0.8.1
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.2
g2p-en==2.1.0
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.71.0
huggingface-hub==0.30.2
hydra-core==1.3.2
idna==3.10
inflect==7.5.0
iniconfig==2.1.0
intervaltree==3.1.0
ipython==8.36.0
ipywidgets==8.1.6
jedi==0.19.2
Jinja2==3.1.6
jiwer==3.1.0
joblib==1.5.0
jupyterlab_widgets==3.0.14
kaldi-python-io==1.2.2
kaldiio==2.18.1
kiwisolver==1.4.8
lazy_loader==0.4
Levenshtein==0.27.1
lhotse==1.31.0
librosa==0.11.0
lightning==2.5.1.post0
lightning-utilities==0.14.3
lilcom==1.8.1
llvmlite==0.44.0
loguru==0.7.3
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
multiprocess==0.70.15
nemo-toolkit==2.2.1
networkx==3.4.2
nltk==3.9.1
numba==0.61.0
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
onnx==1.17.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
plac==1.4.5
platformdirs==4.3.7
pluggy==1.5.0
pooch==1.8.2
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyarrow==20.0.0
pybind11==2.13.6
pycparser==2.22
pydantic==1.10.22
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
pytorch-lightning==2.0.7
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
resampy==0.4.3
rich==14.0.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
sacremoses==0.1.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentencepiece==0.2.0
sentry-sdk==2.27.0
setproctitle==1.3.6
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sortedcontainers==2.4.0
soundfile==0.13.1
sox==1.5.0
soxr==0.5.0.post1
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
text-unidecode==1.3
texterrors==1.0.9
threadpoolctl==3.6.0
tokenizers==0.13.3
tomli==2.2.1
toolz==1.0.0
torch==2.7.0
torchmetrics==1.7.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.28.0
triton==3.3.0
typeguard==4.4.2
typer==0.15.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wandb==0.19.10
wcwidth==0.2.13
webdataset==0.1.62
webrtcvad==2.0.10
Werkzeug==3.1.3
wget==3.2
widgetsnbextension==4.0.14
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
youtokentome==1.0.6
